render light bulb v0

+++ rs_sims/src/mesh/lathe_mesh.rs
@@ -11,7 +11,7 @@ use macroquad::{
 };
 
 /** Return the vertices of the lathe mesh created  */
-fn create_lathe_vertices(profile: &[Vec2], num_rings: u16) -> Vec<Vertex> {
+fn create_lathe_vertices(profile: &[Vec2], num_rings: usize) -> Vec<Vertex> {
     let mut vertices = vec![];
     for ring in 0..num_rings {
         for pt in profile {
@@ -28,14 +28,14 @@ fn create_lathe_vertices(profile: &[Vec2], num_rings: u16) -> Vec<Vertex> {
     vertices
 }
 
-fn create_lathe_indices(profile: &[Vec2], num_rings: u16, vertices: &Vec<Vertex>) -> Vec<u16> {
+fn create_lathe_indices(profile: &[Vec2], num_rings: usize, vertices: &Vec<Vertex>) -> Vec<u16> {
     let mut indices = vec![];
     for ring in 0..num_rings {
         for pt_ind in 0..profile.len() - 1 {
-            let top_right = ring as usize * profile.len() + pt_ind;
-            let bot_right = ring as usize * profile.len() + pt_ind + 1;
+            let top_right = ring * profile.len() + pt_ind;
+            let bot_right = ring * profile.len() + pt_ind + 1;
 
-            let next_ring = ((ring + 1) % num_rings) as usize;
+            let next_ring = (ring + 1) % num_rings;
             let top_left = next_ring * profile.len() + pt_ind;
             let bot_left = next_ring * profile.len() + pt_ind + 1;
 
@@ -63,7 +63,7 @@ fn create_lathe_indices(profile: &[Vec2], num_rings: u16, vertices: &Vec<Vertex>
  * Create a mesh by rotating a 2D `profile` around the Y axis `num_rings` times.
  * The `profile` points are assumed to be clockwise ordered.
  */
-pub fn create_lathe_mesh(profile: &[Vec2], num_rings: u16) -> Mesh {
+pub fn create_lathe_mesh(profile: &[Vec2], num_rings: usize) -> Mesh {
     let vertices = create_lathe_vertices(profile, num_rings);
     let indices = create_lathe_indices(profile, num_rings, &vertices);
     Mesh {
